+++ src/com/maddyhome/idea/vim/helper/Helper.kt
@@ -26,7 +26,7 @@ import java.util.*
 
 /**
  * This annotation is created for test functions (methods).
- * It means that original vim behaviour has small differences from behaviour of IdeaVim.
+ * It means that original vim behavior has small differences from behavior of IdeaVim.
  * [shouldBeFixed] flag indicates whether the given functionality should be fixed
  *   or the given behavior is normal for IdeaVim and should be leaved as is.
  *
@@ -42,19 +42,19 @@ import java.util.*
  *    Hello3
  *
  * Why this annotation exists?
- * After creating some functionality you can understand that IdeaVim has a bit different behaviour, but you
+ * After creating some functionality you can understand that IdeaVim has a bit different behavior, but you
  *   cannot fix it right now because of any reasons (bugs in IDE,
  *   the impossibility of this functionality in IDEA (*[shouldBeFixed] == false*), leak of time for fixing).
  *   In that case, you should NOT remove the corresponding test or leave it without any marks that this test
- *   not fully convenient with vim, but leave the test with IdeaVim's behaviour and put this annotation
+ *   not fully convenient with vim, but leave the test with IdeaVim's behavior and put this annotation
  *   with description of how original vim works.
  *
- * Note that using this annotation should be avoided as much as possible and behaviour of IdeaVim should be as close
+ * Note that using this annotation should be avoided as much as possible and behavior of IdeaVim should be as close
  *   to vim as possible.
  */
 @Retention(AnnotationRetention.SOURCE)
 @Target(AnnotationTarget.FUNCTION)
-annotation class VimBehaviourDiffers(
+annotation class VimBehaviorDiffers(
   val originalVimAfter: String = "",
   val description: String = "",
   val shouldBeFixed: Boolean = true