+++ test/org/jetbrains/plugins/ideavim/action/copy/PutVisualTextMoveCursorActionTest.kt
@@ -23,7 +23,7 @@ import com.maddyhome.idea.vim.VimPlugin
 import com.maddyhome.idea.vim.command.SelectionType
 import com.maddyhome.idea.vim.common.TextRange
 import com.maddyhome.idea.vim.helper.StringHelper.parseKeys
-import com.maddyhome.idea.vim.helper.VimBehaviourDiffers
+import com.maddyhome.idea.vim.helper.VimBehaviorDiffers
 import org.jetbrains.plugins.ideavim.VimTestCase
 import org.junit.Ignore
 
@@ -87,7 +87,7 @@ class PutVisualTextMoveCursorActionTest : VimTestCase() {
     myFixture.checkResult(newFile)
   }
 
-  @VimBehaviourDiffers(originalVimAfter = """
+  @VimBehaviorDiffers(originalVimAfter = """
             A Discovery
 
             ound it in a legendary land