+++ test/org/jetbrains/plugins/ideavim/ex/ExEntryTest.kt
@@ -20,7 +20,7 @@ package org.jetbrains.plugins.ideavim.ex
 
 import com.maddyhome.idea.vim.VimPlugin
 import com.maddyhome.idea.vim.helper.StringHelper
-import com.maddyhome.idea.vim.helper.VimBehaviourDiffers
+import com.maddyhome.idea.vim.helper.VimBehaviorDiffers
 import com.maddyhome.idea.vim.option.OptionsManager
 import com.maddyhome.idea.vim.ui.ExDocument
 import com.maddyhome.idea.vim.ui.ExEntryPanel
@@ -196,7 +196,7 @@ class ExEntryTest : VimTestCase() {
     assertExText("rch")
   }
 
-  @VimBehaviourDiffers(description = "Vim reorders history even when cancelling entry")
+  @VimBehaviorDiffers(description = "Vim reorders history even when cancelling entry")
   fun `test command history`() {
     typeExInput(":set digraph<CR>")
     typeExInput(":digraph<CR>")
@@ -211,7 +211,7 @@ class ExEntryTest : VimTestCase() {
 
     deactivateExEntry()
 
-    // TODO: Vim behaviour reorders the history even when cancelling history
+    // TODO: Vim behavior reorders the history even when cancelling history
 //        typeExInput(":<Up>")
 //        assertExText("set digraph")
 //        typeText("<Up>")
@@ -294,7 +294,7 @@ class ExEntryTest : VimTestCase() {
     assertExText("something cool")
   }
 
-  @VimBehaviourDiffers(description = "Vim reorders history even when cancelling entry")
+  @VimBehaviorDiffers(description = "Vim reorders history even when cancelling entry")
   fun `test matching search history`() {
     typeExInput("/something cool<CR>")
     typeExInput("/not cool<CR>")
@@ -307,7 +307,7 @@ class ExEntryTest : VimTestCase() {
 
     deactivateExEntry()
 
-    // TODO: Vim behaviour reorders the history even when cancelling history
+    // TODO: Vim behavior reorders the history even when cancelling history
 //        typeExInput(":<Up>")
 //        assertEquals("set digraph", exEntryPanel.text)
 //        typeText("<Up>")
@@ -361,7 +361,7 @@ class ExEntryTest : VimTestCase() {
     assertExOffset(0)
   }
 
-  @VimBehaviourDiffers(description = "Moving one word right positions caret at end of previous word")
+  @VimBehaviorDiffers(description = "Moving one word right positions caret at end of previous word")
   fun `test move caret one WORD right`() {
     typeExInput(":set incsearch")
     caret.dot = 0